+++ include/dtp/platform.h
@@ -19,8 +19,8 @@ extern "C"
     typedef struct
     {
         uint32_t size;
-        uint32_t (*read)(uint8_t payload_id, uint32_t offset, void *output, uint32_t size);
-        void (*completed)(uint8_t payload_id, uint32_t size);
+        uint32_t (*read)(uint16_t payload_id, uint32_t offset, void *output, uint32_t size);
+        void (*completed)(uint16_t payload_id, uint32_t size);
     } dtp_payload_meta_t;
 
     /**
