+++ src/payload_api.c
@@ -2,7 +2,7 @@
 #include <dtp/platform.h>
 
 static const char test_data[] = "0123456789ABCDE";
-static uint32_t payload_read(uint8_t payload_id, uint32_t offset, void *output, uint32_t size) {
+static uint32_t payload_read(uint16_t payload_id, uint32_t offset, void *output, uint32_t size) {
     uint32_t res = 0;
     switch(payload_id) {
         case 0:
@@ -27,7 +27,7 @@ static uint32_t payload_read(uint8_t payload_id, uint32_t offset, void *output, 
     return res;
 }
 
-__attribute__((weak)) bool get_payload_meta(dtp_payload_meta_t *meta, uint8_t payload_id) {
+__attribute__((weak)) bool get_payload_meta(dtp_payload_meta_t *meta, uint16_t payload_id) {
     bool result = true;
     switch(payload_id) {
         case 0:
